+++ src/SearchServiceWizard/ISearchServiceWizardContext.ts
@@ -1,0 +1,11 @@
+import { SearchService } from 'azure-arm-search/lib/models';
+import { IResourceGroupWizardContext } from 'vscode-azureextensionui';
+
+export interface ISearchServiceWizardContext extends IResourceGroupWizardContext {
+
+    newServiceName?: string;
+    sku?: string;
+    partitionCount?: number;
+    replicaCount?: number;
+    searchService?: SearchService;
+}