+++ src/SubscriptionTreeItem.ts
@@ -3,14 +3,21 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { SearchManagementClient }  from 'azure-arm-search';
+import { SearchManagementClient, SearchManagementModels }  from 'azure-arm-search';
 import { ResourceManagementClient } from 'azure-arm-resource';
-import { SearchService } from 'azure-arm-search/lib/models';
-import { AzExtTreeItem, createAzureClient, SubscriptionTreeItemBase, addExtensionUserAgent } from 'vscode-azureextensionui';
+import { SearchService, Sku } from 'azure-arm-search/lib/models';
+import { AzExtTreeItem, AzureTreeItem, createAzureClient, SubscriptionTreeItemBase, addExtensionUserAgent,  AzureWizard, AzureWizardPromptStep, ICreateChildImplContext, ILocationWizardContext, ResourceGroupListStep, LocationListStep, StorageAccountCreateStep } from 'vscode-azureextensionui';
 import { SearchServiceTreeItem } from './SearchServiceTreeItem';
 import { getResourcesPath } from "./constants";
 import { Uri } from "vscode";
 import * as path from 'path';
+import { ISearchServiceWizardContext } from './SearchServiceWizard/ISearchServiceWizardContext';
+import { SearchServiceNameStep } from './SearchServiceWizard/SearchServiceNameStep';
+import { nonNullProp } from './utils/nonNull';
+import { SearchServiceSkuStep } from './SearchServiceWizard/SearchServiceSkuStep';
+import { SearchServiceReplicaStep } from './SearchServiceWizard/SearchServiceReplicaStep';
+import { SearchServicePartitionStep } from './SearchServiceWizard/SearchServicePartitionStep';
+import { SearchServiceCreateStep } from './SearchServiceWizard/SearchServiceCreateStep';
 
 export class SubscriptionTreeItem extends SubscriptionTreeItemBase {
     private _nextLink: string | undefined;
@@ -29,7 +36,9 @@ export class SubscriptionTreeItem extends SubscriptionTreeItemBase {
         //      let searchManagementClient = createAzureClient(this.root, SearchManagementClient);
         // but SearchManagementClient defaults to an API version that the Search RP doesn't support. Will get that fixed.
         const searchManagementClient = new SearchManagementClient(this.root.credentials, this.root.subscriptionId, this.root.environment.resourceManagerEndpointUrl);
-        searchManagementClient.apiVersion = "2020-08-01";
+        
+        // This value can't be updated until we upgrade to @azure/arm-search as GET is no longer supported for ListQueryKeys in newer API versions and breaks azure-arm-search
+        searchManagementClient.apiVersion = "2015-08-19";
         addExtensionUserAgent(searchManagementClient);
 
         const resourceManagementClient = createAzureClient(this.root, ResourceManagementClient.ResourceManagementClient);
@@ -49,4 +58,40 @@ export class SubscriptionTreeItem extends SubscriptionTreeItemBase {
     public hasMoreChildrenImpl(): boolean {
         return !!this._nextLink;
     }
+
+    public async createChildImpl(context: ICreateChildImplContext ): Promise<AzureTreeItem> {
+        const searchManagementClient = new SearchManagementClient(this.root.credentials, this.root.subscriptionId, this.root.environment.resourceManagerEndpointUrl);
+        
+        // This value can't be updated until we upgrade to @azure/arm-search as GET is no longer supported for ListQueryKeys in newer API versions and breaks azure-arm-search
+        searchManagementClient.apiVersion = "2015-08-19";
+        addExtensionUserAgent(searchManagementClient);
+
+        const wizardContext: ISearchServiceWizardContext = Object.assign(context, this.root);
+
+        const promptSteps: AzureWizardPromptStep<ISearchServiceWizardContext>[] = [
+            new SearchServiceNameStep(),
+            new ResourceGroupListStep(),
+            new SearchServiceSkuStep(),
+            new SearchServiceReplicaStep(),
+            new SearchServicePartitionStep()
+        ];
+        LocationListStep.addStep(wizardContext, promptSteps);
+
+        const wizard = new AzureWizard(wizardContext, {
+            promptSteps,
+            executeSteps: [new SearchServiceCreateStep()],
+            title: 'Create new Azure Cognitive Search Service'
+        });
+
+        await wizard.prompt();
+
+        const newServiceName: string = nonNullProp(wizardContext, 'newServiceName');
+        context.showCreatingTreeItem(newServiceName);
+        await wizard.execute();
+
+        const newSearchService: SearchService = nonNullProp(wizardContext, 'searchService');
+
+        return new SearchServiceTreeItem(this, newSearchService, searchManagementClient);
+    }
+
 }